feat/add-authentication-for-events-and-users

+++ events-app-backend/src/routes/events.ts
@@ -2,6 +2,7 @@ import { PrismaClient } from '@prisma/client';
 import { Router, Request, Response } from 'express';
 import { Reservation } from '../models/Reservation';
 import { CreateEventInput, Event } from '../models/Event';
+import { authenticate } from '../middlewares/authenticate';
 
 const router = Router();
 const prisma = new PrismaClient();
@@ -21,7 +22,7 @@ router.get('/:id', async (req: Request, res: Response) => {
   res.json(result);
 });
 
-router.post('/', async (req: Request, res: Response) => {
+router.post('/', authenticate, async (req: Request, res: Response) => {
   const { title, description, date, location, capacity, createdBy, price } =
     req.body;
   try {
@@ -44,7 +45,7 @@ router.post('/', async (req: Request, res: Response) => {
   }
 });
 
-router.put('/:id', async (req: Request, res: Response) => {
+router.put('/:id', authenticate, async (req: Request, res: Response) => {
   const { id } = req.params;
   const {
     title,
@@ -56,6 +57,34 @@ router.put('/:id', async (req: Request, res: Response) => {
     price,
   }: CreateEventInput = req.body;
   try {
+    const event = await prisma.event.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    const user = await prisma.user.findUnique({
+      where: {
+        id: parseInt(res.locals.user),
+      },
+    });
+
+    if (!event) {
+      res.status(404).json({ message: 'Event not found' });
+      return;
+    }
+    if (!user) {
+      res.status(404).json({ message: "You're not logged in" });
+      return;
+    }
+
+    if (event?.createdBy !== res.locals.user || res.locals.user !== 'ADMIN') {
+      res
+        .status(403)
+        .json({ message: "Forbidden: you can't access this page" });
+      return;
+    }
+
     const result = await prisma.event.update({
       where: {
         id: parseInt(id),
@@ -76,9 +105,37 @@ router.put('/:id', async (req: Request, res: Response) => {
   }
 });
 
-router.delete('/:id', async (req: Request, res: Response) => {
+router.delete('/:id', authenticate, async (req: Request, res: Response) => {
   const { id } = req.params;
   try {
+    const event = await prisma.event.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    const user = await prisma.user.findUnique({
+      where: {
+        id: parseInt(res.locals.user),
+      },
+    });
+
+    if (!event) {
+      res.status(404).json({ message: 'Event not found' });
+      return;
+    }
+    if (!user) {
+      res.status(404).json({ message: "You're not logged in" });
+      return;
+    }
+
+    if (event?.createdBy !== res.locals.user || res.locals.user !== 'ADMIN') {
+      res
+        .status(403)
+        .json({ message: "Forbidden: you can't access this page" });
+      return;
+    }
+
     const result = await prisma.event.delete({
       where: {
         id: parseInt(id),
